requirements.txt
PySerial
flask
flask-cors
pyserial